Refactor some components to extract some logic
